Moved remaining AST-borrowck diagnostic definitions to `rustc_mir` crate.
